+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+1.2.16 (unreleased)
+-------------------
+
+New:
+
+- *add item here*
+
+Fixes:
+
+- *add item here*
+
+
 1.2.15 (2015-10-28)
 -------------------